refactor: update source code for testing easier

+++ controllers/bookmarkController.py
@@ -9,40 +9,51 @@ class BookmarkController:
     @staticmethod
     def getBookmarkByUserId():
         # NOTE: body contain userId
-        body = request.get_json()
+        userId = request.get_json()['userId']
         try:
-            bookmarks = db.session.query(Bookmark).filter_by(user=body['userId']).all()
+            bookmarks = db.session.query(Bookmark).filter_by(user=userId).all()
             bookmarks = Bookmark.serialize_list(bookmarks)
+            if(not len(bookmarks)):
+                raise
             products = []
             for b in bookmarks:
-                userId = b['user']
                 products.append(Product.query.filter_by(serial_no=b['product']).first().serialize)
             result = {'userID': userId, 'products': products}
             return jsonify(result)
         except:
-            return jsonify({'message': 'The bookmark for userID {} is not existed'.format(body['userId'])}), 400
-
+            return jsonify({'message': 'The bookmark for userID {} is not existed'.format(userId)}), 400
 
     @staticmethod
     def addBookmark():
         # NOTE: body contain userId and serialNo
-        body = request.get_json()
-        bookmark = Bookmark(body['userId'], body['serialNo'])
         try:
-            db.session.add(bookmark)
-            db.session.commit()
+            userId = request.get_json()['userId']
+            serialNo = request.get_json()['serialNo']
+            if(not userId or not serialNo):
+                raise
+            bookmark = Bookmark(userId, serialNo)
+            try:
+                db.session.add(bookmark)
+                db.session.commit()
+            except:
+                return jsonify({'message': 'This bookmark information went wrong'}), 404
+            return jsonify({'message': 'The bookmark is added successfully'})
         except:
-            return jsonify({'message': 'This bookmark information went wrong'}), 404
-        return jsonify({'message': 'The bookmark is added successfully'})
-
+            return jsonify({'message': 'The userId, and serialNo are required'}), 400
     @staticmethod
     def removeBookmark():
         # NOTE: body contain userId and serialNo
-        body = request.get_json()
-        bookmark = db.session.query(Bookmark).filter_by(user=body['userId'], product=body['serialNo']).first()
         try:
-            db.session.delete(bookmark)
-            db.session.commit()
+            userId = request.get_json()['userId']
+            serialNo = request.get_json()['serialNo']
+            if (not userId or not serialNo):
+                raise
+            bookmark = db.session.query(Bookmark).filter_by(user=userId, product=serialNo).first()
+            try:
+                db.session.delete(bookmark)
+                db.session.commit()
+            except:
+                return jsonify({'message': 'This bookmark is not existed'}), 404
+            return jsonify({'message': 'The bookmark is deleted successfully'})
         except:
-            return jsonify({'message': 'This bookmark is not existed'}), 404
-        return jsonify({'message': 'The bookmark is deleted successfully'})
+            return jsonify({'message': 'The userId, and serialNo are required'}), 400
